handle existing files

+++ george/init.py
@@ -2,10 +2,14 @@ import os
 
 def directories():
     for directory in ['code', 'lib', 'test']:
-        os.mkdir(directory)
-        open(os.path.join(directory, '.gitignore'), 'w').write('')
+        if not os.path.isdir(directory):
+            os.mkdir(directory)
+            open(os.path.join(directory, '.gitignore'), 'w').write('')
 
 def readme():
+    if os.path.isfile('readme.md'):
+        return
+
     fp = open('readme.md', 'w')
     for section in ['Overview', 'References', 'Goals', 'Assignment', 'Extra Credit', 'Glossary']:
         fp.write('\n## %s\n' % section)
